asmjist: sync with upstream (nw)

+++ 3rdparty/asmjit/src/asmjit/core/zonevector.h
@@ -36,9 +36,7 @@ ASMJIT_BEGIN_NAMESPACE
 // [asmjit::ZoneVectorBase]
 // ============================================================================
 
-//! \cond INTERNAL
-
-//! Base class implementing core `ZoneVector<>` functionality.
+//! Base class used by \ref ZoneVector template.
 class ZoneVectorBase {
 public:
   ASMJIT_NONCOPYABLE(ZoneVectorBase)
@@ -93,6 +91,7 @@ protected:
   }
 
   //! \}
+  //! \endcond
 
 public:
   //! \name Accessors
@@ -132,8 +131,6 @@ public:
 
   //! \}
 };
-
-//! \endcond
 
 // ============================================================================
 // [asmjit::ZoneVector<T>]
@@ -180,7 +177,7 @@ public:
   inline const T* data() const noexcept { return static_cast<const T*>(_data); }
 
   //! Returns item at the given index `i` (const).
-  inline const T& at(uint32_t i) const noexcept {
+  inline const T& at(size_t i) const noexcept {
     ASMJIT_ASSERT(i < _size);
     return data()[i];
   }
@@ -259,6 +256,7 @@ public:
     return kErrorOk;
   }
 
+  //! Appends `other` vector at the end of this vector.
   inline Error concat(ZoneAllocator* allocator, const ZoneVector<T>& other) noexcept {
     uint32_t size = other._size;
     if (_capacity - _size < size)
@@ -338,6 +336,7 @@ public:
       ::memmove(data, data + 1, size_t(size) * sizeof(T));
   }
 
+  //! Pops the last element from the vector and returns it.
   inline T pop() noexcept {
     ASMJIT_ASSERT(_size > 0);
 
@@ -351,21 +350,33 @@ public:
   }
 
   //! Returns item at index `i`.
-  inline T& operator[](uint32_t i) noexcept {
+  inline T& operator[](size_t i) noexcept {
     ASMJIT_ASSERT(i < _size);
     return data()[i];
   }
 
   //! Returns item at index `i`.
-  inline const T& operator[](uint32_t i) const noexcept {
+  inline const T& operator[](size_t i) const noexcept {
     ASMJIT_ASSERT(i < _size);
     return data()[i];
   }
 
+  //! Returns a reference to the first element of the vector.
+  //!
+  //! \note The vector must have at least one element. Attempting to use
+  //! `first()` on empty vector will trigger an assertion failure in debug
+  //! builds.
   inline T& first() noexcept { return operator[](0); }
+  //! \overload
   inline const T& first() const noexcept { return operator[](0); }
 
+  //! Returns a reference to the last element of the vector.
+  //!
+  //! \note The vector must have at least one element. Attempting to use
+  //! `last()` on empty vector will trigger an assertion failure in debug
+  //! builds.
   inline T& last() noexcept { return operator[](_size - 1); }
+  //! \overload
   inline const T& last() const noexcept { return operator[](_size - 1); }
 
   //! \}
@@ -408,6 +419,7 @@ public:
 // [asmjit::ZoneBitVector]
 // ============================================================================
 
+//! Zone-allocated bit vector.
 class ZoneBitVector {
 public:
   typedef Support::BitWord BitWord;
@@ -689,7 +701,6 @@ public:
   };
 
   //! \}
-
 };
 
 //! \}
